// Apigateway/src/feedback/comments/dto/CommentResponse.interface.ts
export interface ImageDTO {
    _id: string;
    IdUser: string;
    data: string;
    __v: number;
  }
  
  export interface EntrepriseData {
     IdUser: string;
     nameE: string;
     activitySectorE: string;
     categoryE: string;
     titleEntreprise: string;
     sizeE: string;
     descriptionE: string;
     countryE: string;
     regionE: string;
     villeE: string;
     ZIPcodeE: string;
     adresse: string;
     streetNumber: string;
     telE: string[];
     telOwner: string[];
     linkWebSite: string;
  }
  

  
  export interface CommentResponse{
  
    IdUser: string;
    entrepriseId: string;
    //consultantId: String;
    likes: number;
    content: string;
    likedBy: string[]; 
    commentedAt: Date;
    status: string;
    entrepriseimages: ImageDTO[];
    entreprisedata: EntrepriseData[];
  }